extraced receive function from tcpserver
